[INTERNAL] Extract compile from index.js
