+++ lib/index.js
@@ -1,21 +1,17 @@
 "use strict";
 
 const path = require("path");
-const clone = require("clone");
 const css = require("css");
 
-const createParser = require("./less/parser");
+
 const diff = require("./diff");
 const scope = require("./scope");
 
 const fileUtilsFactory = require("./fileUtils");
 
-const themingParametersDataUri = require("./themingParameters/dataUri");
+const Compiler = require("./Compiler");
 
-// Plugins
-const ImportCollectorPlugin = require("./plugin/import-collector");
-const VariableCollectorPlugin = require("./plugin/variable-collector");
-const UrlCollector = require("./plugin/url-collector");
+const themingParametersDataUri = require("./themingParameters/dataUri");
 
 // Workaround for a performance issue in the "css" parser module when used in combination
 // with the "colors" module that enhances the String prototype.
@@ -117,10 +113,6 @@ Builder.prototype.cacheTheme = function(result) {
 Builder.prototype.build = function(options) {
 	const that = this;
 
-	// Stores mapping between "virtual" paths (used within LESS) and real filesystem paths.
-	// Only relevant when using the "rootPaths" option.
-	const mFileMappings = {};
-
 	// Assign default options
 	options = Object.assign({
 		lessInput: null,
@@ -145,172 +137,11 @@ Builder.prototype.build = function(options) {
 		options.parser.relativeUrls = true;
 	}
 
-	function fileHandler(file, currentFileInfo, handleDataAndCallCallback, callback) {
-		let pathname;
-
-		// support absolute paths such as "/resources/my/base.less"
-		if (path.posix.isAbsolute(file)) {
-			pathname = path.posix.normalize(file);
-		} else {
-			pathname = path.posix.join(currentFileInfo.currentDirectory, file);
-		}
-
-		that.fileUtils.readFile(pathname, options.rootPaths).then(function(result) {
-			if (!result) {
-				// eslint-disable-next-line no-console
-				console.log("File not found: " + pathname);
-				callback({type: "File", message: "Could not find file at path '" + pathname + "'"});
-			} else {
-				try {
-					// Save import mapping to calculate full import paths later on
-					mFileMappings[currentFileInfo.rootFilename] = mFileMappings[currentFileInfo.rootFilename] || {};
-					mFileMappings[currentFileInfo.rootFilename][pathname] = result.path;
-
-					handleDataAndCallCallback(pathname, result.content);
-				} catch (e) {
-					// eslint-disable-next-line no-console
-					console.log(e);
-					callback(e);
-				}
-			}
-		}, function(err) {
-			// eslint-disable-next-line no-console
-			console.log(err);
-			callback(err);
-		});
-	}
-
-	async function compile(config) {
-		const parserOptions = clone(options.parser);
-		let rootpath;
-
-		if (config.path) {
-			// Keep rootpath for later usage in the ImportCollectorPlugin
-			rootpath = config.path;
-			parserOptions.filename = config.localPath;
-		}
-
-		// inject the library name as prefix (e.g. "sap.m" as "_sap_m")
-		if (options.library && typeof options.library.name === "string") {
-			const libName = config.libName = options.library.name;
-			config.libPath = libName.replace(/\./g, "/");
-			config.prefix = "_" + libName.replace(/\./g, "_") + "_";
-		} else {
-			config.prefix = ""; // no library, no prefix
-		}
-
-		// Keep filename for later usage (see ImportCollectorPlugin) as less modifies the parserOptions.filename
-		const filename = parserOptions.filename;
-
-		// Only use custom file handler when "rootPaths" or custom "fs" are used
-		let fnFileHandler;
-		if ((options.rootPaths && options.rootPaths.length > 0) || that.customFs) {
-			fnFileHandler = fileHandler;
-		}
-
-		const parser = createParser(parserOptions, fnFileHandler);
-
-		function parseContent(content) {
-			return new Promise(function(resolve, reject) {
-				parser.parse(content, function(err, tree) {
-					if (err) {
-						reject(err);
-					} else {
-						resolve(tree);
-					}
-				});
-			});
-		}
-
-		const tree = await parseContent(config.content);
-		const result = {tree};
-
-		// plugins to collect imported files and variable values
-		const oImportCollector = new ImportCollectorPlugin({
-			importMappings: mFileMappings[filename]
-		});
-		const oVariableCollector = new VariableCollectorPlugin(options.compiler);
-		const oUrlCollector = new UrlCollector();
-
-		// render to css
-		result.css = tree.toCSS(Object.assign({}, options.compiler, {
-			plugins: [oImportCollector, oVariableCollector, oUrlCollector]
-		}));
-
-		// retrieve imported files
-		result.imports = oImportCollector.getImports();
-
-		// retrieve reduced set of variables
-		result.variables = oVariableCollector.getVariables(Object.keys(mFileMappings[filename] || {}));
-
-		// retrieve all variables
-		result.allVariables = oVariableCollector.getAllVariables();
-
-		// also compile rtl-version if requested
-		let oRTL;
-		if (options.rtl) {
-			const RTLPlugin = require("./plugin/rtl");
-			oRTL = new RTLPlugin();
-
-			const urls = oUrlCollector.getUrls();
-
-			const existingImgRtlUrls = (await Promise.all(
-				urls.map(async ({currentDirectory, relativeUrl}) => {
-					const relativeImgRtlUrl = RTLPlugin.getRtlImgUrl(relativeUrl);
-					if (relativeImgRtlUrl) {
-						const resolvedImgRtlUrl = path.posix.join(currentDirectory, relativeImgRtlUrl);
-						if (await that.fileUtils.findFile(resolvedImgRtlUrl, options.rootPaths)) {
-							return resolvedImgRtlUrl;
-						}
-					}
-				})
-			)).filter(Boolean);
-
-			oRTL.setExistingImgRtlPaths(existingImgRtlUrls);
-		}
-
-		if (oRTL) {
-			result.cssRtl = tree.toCSS(Object.assign({}, options.compiler, {
-				plugins: [oRTL]
-			}));
-		}
-
-		if (rootpath) {
-			result.imports.unshift(rootpath);
-		}
-
-		// also compile css-variables version if requested
-		if (options.cssVariables) {
-			// parse the content again to have a clean tree
-			const cssVariablesSkeletonTree = await parseContent(config.content);
-
-			// generate the skeleton-css and the less-variables
-			const CSSVariablesCollectorPlugin = require("./plugin/css-variables-collector");
-			const oCSSVariablesCollector = new CSSVariablesCollectorPlugin(config);
-			const oVariableCollector = new VariableCollectorPlugin(options.compiler);
-			result.cssSkeleton = cssVariablesSkeletonTree.toCSS(Object.assign({}, options.compiler, {
-				plugins: [oCSSVariablesCollector, oVariableCollector]
-			}));
-			const varsOverride = oVariableCollector.getAllVariables();
-			result.cssVariablesSource = oCSSVariablesCollector.toLessVariables(varsOverride);
-
-			if (oRTL) {
-				const oCSSVariablesCollectorRTL = new CSSVariablesCollectorPlugin(config);
-				result.cssSkeletonRtl = cssVariablesSkeletonTree.toCSS(Object.assign({}, options.compiler, {
-					plugins: [oCSSVariablesCollectorRTL, oRTL]
-				}));
-			}
-
-			// generate the css-variables content out of the less-variables
-			const cssVariablesTree = await parseContent(result.cssVariablesSource);
-			const CSSVariablesPointerPlugin = require("./plugin/css-variables-pointer");
-			result.cssVariables = cssVariablesTree.toCSS(Object.assign({}, options.compiler, {
-				plugins: [new CSSVariablesPointerPlugin()]
-			}));
-		}
-
-		return result;
-	}
+	const compiler = new Compiler({
+		options,
+		fileUtils: this.fileUtils,
+		customFs: this.customFs
+	});
 
 	async function addInlineParameters(result) {
 		return themingParametersDataUri.addInlineParameters({result, options});
@@ -346,13 +177,13 @@ Builder.prototype.build = function(options) {
 				if (!config) {
 					throw new Error("Could not find embeddedCompareFile at path '" + scopeOptions.embeddedCompareFilePath + "'");
 				}
-				return compile(config);
+				return compiler.compile(config);
 			}),
 			that.fileUtils.readFile(scopeOptions.embeddedFilePath, options.rootPaths).then(function(config) {
 				if (!config) {
 					throw new Error("Could not find embeddedFile at path '" + scopeOptions.embeddedFilePath + "'");
 				}
-				return compile(config);
+				return compiler.compile(config);
 			})
 		]).then(function(results) {
 			return {
@@ -494,7 +325,9 @@ Builder.prototype.build = function(options) {
 			}
 
 			// No css diffing and scoping
-			return that.fileUtils.readFile(options.lessInputPath, options.rootPaths).then(compile);
+			return that.fileUtils.readFile(options.lessInputPath, options.rootPaths).then((config) => {
+				return compiler.compile(config);
+			});
 		});
 	}
 
@@ -507,7 +340,7 @@ Builder.prototype.build = function(options) {
 	}
 
 	if (options.lessInput) {
-		return compile({
+		return compiler.compile({
 			content: options.lessInput
 		}).then(addInlineParameters).then(that.cacheTheme.bind(that));
 	} else {
